Do not Use gtk_widget_override_background_color for overide background color,use CSS for it Possible Fix Memory on Downloads Some Logical Fixies around active/pasive in ADC

+++ dcpp/ClientManager.h
@@ -153,8 +153,8 @@ public:
 				ip = aIpPort;
 			} else {
 				ip = aIpPort.substr(0, i);
-				port = Util::toInt(aIpPort.substr(i + 1));
-			}	
+				port = (uint16_t)Util::toInt(aIpPort.substr(i + 1));
+			}
 	}
 
 private:
